+++ editor/app/(demo)/demo/sales/campaign/polestar-kr-2503/[cid]/invite/page.tsx
@@ -1,19 +1,15 @@
-"use client";
-
 import React from "react";
 import { ScreenWindowRoot } from "@/theme/templates/kit/components";
 import TMP_Invite from "./component";
 
-export default function Page({
-  params,
-}: {
-  params: {
-    cid: string;
-  };
-}) {
+type Params = {
+  cid: string;
+};
+
+export default async function Page({ params }: { params: Promise<Params> }) {
   return (
     <ScreenWindowRoot>
-      <TMP_Invite params={params} />
+      <TMP_Invite params={await params} />
     </ScreenWindowRoot>
   );
 }
