+++ editor/app/(dev)/playground/[slug]/page.tsx
@@ -1,18 +1,20 @@
 import { createServerComponentClient } from "@/lib/supabase/server";
 import { cookies } from "next/headers";
-import { notFound, redirect } from "next/navigation";
+import { redirect } from "next/navigation";
 import Playground from "@/scaffolds/playground";
 
 export const maxDuration = 60;
 
+type Params = {
+  slug: string;
+};
+
 export default async function SharedPlaygroundPage({
   params,
 }: {
-  params: {
-    slug: string;
-  };
+  params: Promise<Params>;
 }) {
-  const { slug } = params;
+  const { slug } = await params;
   const cookieStore = await cookies();
   const supabase = createServerComponentClient(cookieStore);
