+++ editor/app/(e)/d/e/[id]/badrequest/page.tsx
@@ -9,12 +9,14 @@ import Link from "next/link";
 import i18next from "i18next";
 import { ssr_page_init_i18n } from "@/i18n/ssr";
 
+type Params = { id: string };
+
 export default async function BadRequestPage({
   params,
 }: {
-  params: { id: string };
+  params: Promise<Params>;
 }) {
-  const form_id = params.id;
+  const { id: form_id } = await params;
   await ssr_page_init_i18n({ form_id });
 
   return (
