+++ editor/app/(site)/organizations/[organization_name]/people/page.tsx
@@ -5,14 +5,14 @@ import { notFound, redirect } from "next/navigation";
 import { PublicUrls } from "@/services/public-urls";
 import MemberList, { MemberItem } from "./list";
 
+type Params = { organization_name: string };
+
 export default async function PoeplesPage({
   params,
 }: {
-  params: {
-    organization_name: string;
-  };
+  params: Promise<Params>;
 }) {
-  const organization_name = params.organization_name;
+  const { organization_name } = await params;
   const cookieStore = await cookies();
   const supabase = createRouteHandlerWorkspaceClient(cookieStore);
   const avatar_url = PublicUrls.organization_avatar_url(supabase);
