+++ editor/app/(workspace)/[org]/[proj]/page.tsx
@@ -24,6 +24,7 @@ export default function FormsDashboardPage({
   params,
   searchParams,
 }: {
+  // TODO: [next15](https://nextjs.org/docs/app/building-your-application/upgrading/version-15#asynchronous-page)
   params: {
     org: string;
     proj: string;
